Struct hash_rings::jump::Ring

pub struct Ring<H = RandomState> { /* fields omitted */ }

A hashing ring implemented using jump hashing.

Jump hashing is based on using a hash of the key as the seed for a random number generator and using it to jump forward in a list of nodes until it falls off the end. The last node it lands on is the result.

Jump hashing is very fast and executes in O(ln n) time. It also has no memory overhead and has virtually perfect key distribution. However, the main limitation of jump hashing is that it returns an integer in the range [0, nodes) and it does not support arbitrary node names.

Examples

use hash_rings::jump::Ring;
use std::collections::hash_map::DefaultHasher;
use std::hash::BuildHasherDefault;

type DefaultBuildHasher = BuildHasherDefault<DefaultHasher>;

let ring = Ring::with_hasher(DefaultBuildHasher::default(), 100);

assert_eq!(ring.get_node(&"foo"), 8);
assert_eq!(ring.nodes(), 100);

Methods

impl Ring<RandomState>

pub fn new(nodes: u32) -> Self

Constructs a new Ring with a specified number of nodes.

Panics

Panics if the number of nodes is zero.

Examples

use hash_rings::jump::Ring;

let ring: Ring = Ring::new(100);

impl<H> Ring<H>

pub fn with_hasher(hash_builder: H, nodes: u32) -> Self

Constructs a new Ring with a specified number of nodes and hash builder.

Panics

Panics if the number of nodes is zero.

Examples

use hash_rings::jump::Ring;
use std::collections::hash_map::DefaultHasher;
use std::hash::BuildHasherDefault;

type DefaultBuildHasher = BuildHasherDefault<DefaultHasher>;

let ring: Ring<_> = Ring::with_hasher(DefaultBuildHasher::default(), 100);

pub fn get_node<T>(&self, key: &T) -> u32 where
T: Hash,
H: BuildHasher,

Returns the node associated with a key.

Examples

use hash_rings::jump::Ring;
use std::collections::hash_map::DefaultHasher;
use std::hash::BuildHasherDefault;

type DefaultBuildHasher = BuildHasherDefault<DefaultHasher>;

let ring = Ring::with_hasher(DefaultBuildHasher::default(), 100);
assert_eq!(ring.get_node(&"foo"), 8);

pub fn nodes(&self) -> u32

Returns the number of nodes in the ring.

Examples

use hash_rings::jump::Ring;

let ring = Ring::new(100);
assert_eq!(ring.nodes(), 100);